--- webapp/src/components/NoData.js
@@ -1,7 +1,0 @@
-
-import React from 'react';
-
-// Text displayed when there's no data available
-export default function NoData() {
-  return <div className="no-data">No data available.</div>;
-}
