+++ webapp/src/components/Publisher.js
@@ -13,6 +13,12 @@ import {
 } from "../utils/data";
 import {NumberCell, PercentCell, TableHeaderTip} from "./Table";
 import {ActiveHeadingComponent, OpenAccessHeadingComponent} from "./PublisherTable";
+import {Chart} from "./chart";
+import {NoData, AccessLogs, TopArticlesTable} from "./details";
+
+import coverageSpec from './publisher-coverage-chart.json';
+import quantileSpec from './journal-quantile-chart.json';
+
 
 export default class PublisherDetails extends React.Component {
   constructor(props) {
@@ -51,8 +57,64 @@ export default class PublisherDetails extends React.Component {
           <PublisherOverallCoverage data={this.state.info} />
         </div>
 
+        <div className={styles.section}>
+          <h3 className="">Yearly coverage</h3>
+          {this._coverageChart()}
+        </div>
 
+        {this.state.info.access_logs && <div className={styles.section}>
+          <h3 className="">Access logs summary</h3>
+          <p className="section-description">Averages for how often articles published from January 2014 to October 2015 were accessed via Sci-Hub during September 2015 through February 2016.</p>
+          <AccessLogs data={this.state.info.access_logs}/>
+        </div>}
+
+        <div className={styles.section}>
+          <h3 className="">Access logs visitor distribution</h3>
+          <p className="section-description">Visitor percentiles are plotted for articles published from January 2014 to October 2015. Visitors refers to the number of unique IP addressed that accessed the article via Sci-Hub from September 2015 to February 2016.</p>
+          {this._quantilesChart()}
+        </div>
+
+        <div className={styles.section}>
+          <h3 className="">Access logs top articles</h3>
+          <p className="section-description">The following table shows the 100 most visited articles in Sci-Hub's access logs from September 2015 through February 2016.</p>
+          {this._topArticles()}
+        </div>
       </div>;
+    }
+  }
+
+  _topArticles() {
+    if (this.state.articles) {
+      return <TopArticlesTable data={this.state.articles} journal={true}/>;
+    } else {
+      return <NoData/>;
+    }
+  }
+
+  _quantilesChart() {
+    if (this.state.quantiles) {
+      return <Chart spec={quantileSpec} data={this.state.quantiles} />;
+    } else {
+      return <NoData/>;
+    }
+  }
+
+  _coverageChart() {
+    if (this.state.coverage) {
+      return <Chart spec={coverageSpec} data={this.state.coverage} adjustWidth={-105}
+                    tooltip={ {
+                      showAllFields: false,
+                      fields: [
+                        {field: 'scihub', title: 'Sci-Hub', formatType: "number", format: ',d'},
+                        {field: 'scihub_journals', title: 'Sci-Hub Journals', formatType: "number", format: ',d'},
+                        {field: 'crossref', title: 'Crossref', formatType: "number", format: ',d'},
+                        {field: 'crossref_journals', title: 'Crossref Journals', formatType: "number", format: ',d'},
+                        {field: 'coverage', title: 'Coverage', formatType: "number", format: ',.1%'},
+                        {field: 'year', title: 'Year', formatType: 'time', format: '%Y'},
+                      ]
+                    } }/>;
+    } else {
+      return <NoData/>;
     }
   }
 }
@@ -67,8 +129,8 @@ function PublisherOverallCoverage({data}) {
           ['Crossref', 'The total number of articles in journals from the publisher, as extracted from Crossref.'],
           ['Coverage', 'The number of articles in Sci-Hub divided by the total number of articles.'],
          ].map(([title, tooltip], index) => <th key={index} className="griddle-table-heading-cell">{title} <TableHeaderTip tooltip={tooltip}/></th>)}
-        <th className="griddle-table-heading-cell"><OpenAccessHeadingComponent /></th>
-        <th className="griddle-table-heading-cell"><ActiveHeadingComponent /></th>
+        <th className="griddle-table-heading-cell"><OpenAccessHeadingComponent className=""/></th>
+        <th className="griddle-table-heading-cell"><ActiveHeadingComponent className=""/></th>
       </tr>
       </thead>
       <tbody className="griddle-table-body">
@@ -84,5 +146,4 @@ function PublisherOverallCoverage({data}) {
     </table>
   </div>
 }
-
