+++ webapp/src/components/PublisherTable.js
@@ -64,8 +64,8 @@ PublisherCell = connect((state, props) => ({
   rowData: rowDataSelector(state, props)
 }))(PublisherCell);
 
-export const OpenAccessHeadingComponent = ({icon}) =>
-  <div className="text-center">
+export const OpenAccessHeadingComponent = ({icon, className = 'text-center'}) =>
+  <div className={className}>
     <a className={tableStyles.header} href="javascript:void(0)">
       <Tooltip title="The percent of the publisher's articles that are in open access journals. Note that open access articles in hybrid journals do not count towards this measure">
         <img src="https://upload.wikimedia.org/wikipedia/commons/7/77/Open_Access_logo_PLoS_transparent.svg"
@@ -75,8 +75,8 @@ export const OpenAccessHeadingComponent = ({icon}) =>
     </a>
   </div>;
 
-export const ActiveHeadingComponent = ({icon}) =>
-  <div className="text-center">
+export const ActiveHeadingComponent = ({icon, className = 'text-center'}) =>
+  <div className={className}>
     <a className={tableStyles.header} href="javascript:void(0)">
       <Tooltip title="The percent of the publisher's articles that are in active journals (journals that still publish new articles)">
         <i className="glyphicon glyphicon-ok text-success"/>
